Show full day name in calendar date display
Changed from "Tue, 4 Feb 2026" to "Tuesday, 4 Feb 2026"

+++ app/[slug]/admin/calendar/page.tsx
@@ -421,7 +421,7 @@ export default function CalendarPage() {
   function formatDateDisplay(dateStr: string) {
     const [year, month, day] = dateStr.split('-').map(Number);
     const date = new Date(year, month - 1, day);
-    return date.toLocaleDateString("en-GB", { weekday: "short", day: "numeric", month: "short", year: "numeric" });
+    return date.toLocaleDateString("en-GB", { weekday: "long", day: "numeric", month: "short", year: "numeric" });
   }
 
   function openAppointment(apt: Appointment) {
